Add files via upload
Ping-Pong, is an indoor game played on a flat table between two players. The objective of the game is to hit the ball over the net and bounce it to the other side of the table in a way that the opponent misses hitting it back. This game is trendy across the world. A multiplayer Game which is made using html, css and javascript